update 10 fen file

+++ four_week/4_100_8_fanshu.cpp
@@ -3,10 +3,10 @@
 using namespace std;
 int main() {
 	int n;
-	int s=0;
+	double s=0;
 	cin>>n;
 	double re=0.0;
-	int num[n];
+	double num[n];
 	for(int i=0;i<n;i++){
 		cin>>num[i];
 	}
